[CoreBundle] Workspace copy refactor (using serializers) (#5711)
* [CoreBundle] Resource cleanup.

* Removig text manager.

* Directory copy simplify.

* Removing more code.

* move code

* Removing a lot of useless code.

* more code removed

* more simplify

* Fix notifications.

* ok

* blog listener fix.

* Travis

* travis

* base

* wip

* wip

* [CoreBundle] Widgets Serializers.

* Fix #5688 locally.

* changing bool

* Simplify

* entities refresh

* Removing alias

* nofetch

* Options

* J'arrive pas

* MessageSerializer.

* UserSerializer

* RoleSerializer

* [CoreBundle] GroupSerializer.

* ScheduledTaskSerializer.

* ResouceCommentSerialiezr

* WorkspaceSerializer

* FacetSerializer

* wip

* FacetSerializers

* Agenda

* Competency Serializer

* ForumSerializer

* Lesson.

* LTI

* Badges

* PlannedNotification.

* test

* Some specific serializers.

* remove var dump

* Requires testing but might be ok.

* First draft

* fix ws creation

* Travis

* Using serializer now.

* Fixing install error.

* wip

* wip

* WikiSerializer

* test

* wip

* test

* Import working again + simplifications

* Widget uuid refresh + default workspace.

* Fixing creation

* Workspace copy doesnt duplicate path resources anymore !

* blog copy & widgets

* wip

* Claco form properly copied (I think)

* Move some code.

* Install

* Keep fixing stuff.

* Forum

* [CoreBundle] Copy from controller.

* wip

* Travis

* will it work ?

* more fixes

* Travis

+++ plugin/claco-form/Serializer/ClacoFormSerializer.php
@@ -195,7 +195,7 @@ class ClacoFormSerializer
      *
      * @return ClacoForm
      */
-    public function deserialize($data, ClacoForm $clacoForm)
+    public function deserialize($data, ClacoForm $clacoForm, array $options = [])
     {
         // TODO : remove and call all setters individually
         $this->sipe('details', 'setDetails', $data, $clacoForm);
@@ -224,30 +224,32 @@ class ClacoFormSerializer
         $newFieldsUuids = [];
         $clacoForm->emptyFields();
 
-        foreach ($data['fields'] as $fieldData) {
-            if (isset($fieldData['id'])) {
-                $newFieldsUuids[] = $fieldData['id'];
-            }
-            $field = isset($fieldData['id']) ? $this->fieldRepo->findOneBy(['uuid' => $fieldData['id']]) : null;
+        if (!in_array(Options::REFRESH_UUID, $options)) {
+            foreach ($data['fields'] as $fieldData) {
+                if (isset($fieldData['id'])) {
+                    $newFieldsUuids[] = $fieldData['id'];
+                }
+                $field = isset($fieldData['id']) ? $this->fieldRepo->findOneBy(['uuid' => $fieldData['id']]) : null;
 
-            if (empty($field)) {
-                $field = new Field();
-                $field->setClacoForm($clacoForm);
-            }
-            $newField = $this->fieldSerializer->deserialize($fieldData, $field);
-            $this->om->persist($newField);
+                if (empty($field)) {
+                    $field = new Field();
+                    $field->setClacoForm($clacoForm);
+                }
+                $newField = $this->fieldSerializer->deserialize($fieldData, $field);
+                $this->om->persist($newField);
 
-            $clacoForm->addField($newField);
+                $clacoForm->addField($newField);
+            }
+            $this->om->startFlushSuite();
+
+            /* Removes previous fields that are not used anymore */
+            foreach ($oldFields as $field) {
+                if (!in_array($field->getUuid(), $newFieldsUuids)) {
+                    $this->deleteField($field);
+                }
+            }
+            $this->om->endFlushSuite();
         }
-        $this->om->startFlushSuite();
-
-        /* Removes previous fields that are not used anymore */
-        foreach ($oldFields as $field) {
-            if (!in_array($field->getUuid(), $newFieldsUuids)) {
-                $this->deleteField($field);
-            }
-        }
-        $this->om->endFlushSuite();
 
         // entry list config
         // todo : big c/c from Claroline\CoreBundle\API\Serializer\Widget\Type\ListWidgetSerializer
